Refactor LiveView helpers (#322)
